refactor: move remote mutation locks into Deck entity
