+++ src/entities/deck/@x/card.ts
@@ -2,3 +2,4 @@ import type { Deck } from "../model/deck";
 
 export type { DeckId } from "../model/deck";
 export type DeckForCard = Pick<Deck, "id" | "uid">;
+export { deckMembershipMutationLock, withDeckMembershipLocks } from "../api/remoteMutationLocks";
